Adding the Authenticator Plug
Setting up the current_user in the 'conn' is now done
correctly by creating the Authenticator Plug that hooks
into the router :browser pipeline where it get the
current_user (nil if not authenticated) and stores it
in the 'conn' object.

Hence, current_user will always be there (possibly nil when
not authenticted) when calling for rendering from the controllers.