Adding schema associations and corresponding migrations
It was (is?) quite confusing how to set upp the
associations suck as belongs_to, has_many and has_one.
Especially the migrations, so I'm not sure this is correct.

It seems like it is important to number the migration files
so that they will run in order, where oldest runs first or
e.g if a field can't be added until another table has been created.